refactored core.cs to see how it feels
